+++ syn-grammar/tests/rule_arguments_test.rs
@@ -6,7 +6,7 @@ use syn_grammar::testing::Testable;
 fn test_rule_arguments() {
     grammar! {
         grammar args {
-            pub rule main -> i32 = "start" v:value<_>(10) -> { v }
+            pub rule main -> i32 = "start" v:value(offset=10) -> { v }
             rule value(offset: i32) -> i32 = i:i32 -> { i + offset }
         }
     }